Sometime ctrl+c doesn't close the program

+++ include/ServerConf.hpp
@@ -24,7 +24,7 @@ class ServerConf {
 		
 		// getter member
 		std::string	getHost();
-		size_t		getPort();
+		u_int16_t	getPort();
 		std::string	getServerName();
 
 		// member function
